+++ security/yara/YaraEngine_new.py
